+++ main.py
@@ -1,4 +1,4 @@
-from automater.runner import main
+from twrar.runner import main
 
 if __name__ == "__main__":
     main()